Firm full created successfully

+++ database/migrations/2022_09_25_052028_create_firms_table.php
@@ -16,9 +16,9 @@ return new class extends Migration
         Schema::create('firms', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('all_sum');
-            $table->integer('indebtedness');
-            $table->integer('given_sum');
+            $table->unsignedBigInteger('all_sum');
+            $table->unsignedBigInteger('indebtedness');
+            $table->unsignedBigInteger('given_sum');
             $table->timestamps();
             $table->softDeletes();
         });
